Refactored: Moved useTheme() hook
