+++ src/client/plugin/contexts/themeContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext } from 'react';
+import { createContext } from 'react';
 import { ColorTheme } from '../../lib/types';
 
 interface ThemeContextType {
@@ -9,14 +9,6 @@ interface ThemeContextType {
   setIsSettingsOpen: (open: boolean) => void;
 }
 
-export const ThemeContext = createContext<ThemeContextType | undefined>(
-  undefined
-);
+const themeContext = createContext<ThemeContextType | undefined>(undefined);
 
-export function useTheme() {
-  const context = useContext(ThemeContext);
-  if (context === undefined) {
-    throw new Error('useTheme must be used within a ThemeProvider');
-  }
-  return context;
-}
+export default themeContext;
